precompute probabilities: generate two single level dictionaries.
This turns the d_graph structure into `probabilities` and `neighbors`:
the first maps `id` and `(id, id)` to an array of transition
probabilities (single `id` is first-travel probabilities, pair is
`(source, current)` probabilities for a walk ending "..., source,
current"), and `neighbors` maps `id` to a list of neighbors.

This is a relatively small single-threaded performance improvement,
but is crucial to break the interloop dependencies so that this can be
parallelised.

Computing transition probabilities on a graph with 10000 nodes and
average degree 20 goes from 1850 it/s to 1950 it/s (5% faster).