use the alias method for neighbor sampling.
This makes computing the probabilities slower (for a graph with 10000
and an average degree, it goes from ~1900it/s to ~700it/s), but doing
the walks are significantly faster, meaning that num_walks=10,
walk_length=20 on that graph goes from 1370it/s to 5400it/s. Even with
the extra overhead of encoding/decoding more (complicated) data, the
overall time for the Node2Vec constructor with the above
parameters (with one worker) goes from 75s to 32s.

The example.py script also goes from 14.5s real (32.3s user) to 10.2s
real (15.1s user), indicating 17.2s of CPU time was cut off (which
matches the 4.3s of real time spread across 4 CPUs).