generate walks: swap num_walks and neighbors loops.
This allows moving more loop-invariant code out of the inner loops,
including moving the randomnness generation slightly further
out (don't want to move it all the way, because preallocating for all
the nodes ends up taking a lot of memory).

The iteration time for walk_length=20, num_walks=10 on the 10k graph
with average degree 20 goes from 13600 it/s to 14500 it/s (7% faster).